Add functionality to extract data files
This tool will eventually be able to do many things, so we'll prepare for just that. The change here introduces a new extract mode and command line interface. The former allows us to ignore QuickBMS entirely, since we can do the whole extracting process several times faster.

On my machine, this new mode is done in under 5 minutes. QuickBMS for the same content requires about 29 minutes with the most optimal script I could come up with. As a massive benefit, unlike QuickBMS, we don't randomly have 8 bytes from a completely random address infecting our exported files.

For now this disables/removes the whole meshinfo/mesh stuff. We're pretty sure that the whole thing is a single file now anyway, so we need to implement specific modes for that now.