Update .texture 010 template with discovered information
We've mostly discovered everything useful now, and can use this to fully decode the file. It's still unclear why some of the .texture data resides in gpu_resources instead of stream, but it is clear now why the DDS header resides in the main file.

Technically, this information is enough to write a texture exporter now.