#1: Figured out how the node list structure works
Tough nut to crack, but I've finally made progress. Turns out it wasn't two 4x4 matrices, but instead one structure containing the data to make the matrix, and an actual list of precalculated matrices. This could have been used to offset the abysmal performance of the AMD Bulldozer CPUs/APUs that were in use around the time the engine was alive and maintained.